pyproject.toml
[tool.poetry]
name = "dz-dj-1"
version = "0.1.0"
description = "Первое домашнее задание по Django."
authors = ["Павел <chernovpp@gmail.com>"]
readme = "README.md"
packages = [{include = "dz_dj_1"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
pillow = "^9.4.0"
psycopg2 = "^2.9.5"
django-extensions = "^3.2.1"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-factoryboy = "^2.5.1"
pytest-django = "^4.5.2"
pylint = "^2.16.2"
pylint-django = "^2.5.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
